feat: update OSKPTile in activity
