+++ packages/cryptoplease/lib/features/outgoing_split_key_payments/bl/repository.dart
@@ -61,7 +61,7 @@ enum OSKPStatusDto {
   txLinksFailure,
 }
 
-extension on OSKPRow {
+extension OSKPRowExt on OSKPRow {
   Future<OutgoingSplitKeyPayment> toModel(TokenList tokens) async =>
       OutgoingSplitKeyPayment(
         id: id,
