+++ packages/cryptoplease/lib/features/pending_activities/components/styles.dart
@@ -1,0 +1,15 @@
+import 'package:cryptoplease_ui/cryptoplease_ui.dart';
+import 'package:flutter/material.dart';
+
+const titleStyle = TextStyle(
+  fontSize: 16,
+  letterSpacing: .23,
+  color: CpColors.menuPrimaryTextColor,
+  fontWeight: FontWeight.w500,
+);
+
+const subtitleStyle = TextStyle(
+  fontSize: 14,
+  color: CpColors.menuPrimaryTextColor,
+  letterSpacing: .19,
+);
